requirements.txt
discord.py>=2.3
python-dotenv>=1.0
google-generativeai>=0.3.0
aiofiles>=23.0.0
asyncio>=3.4.3
python-dateutil>=2.8.2
structlog>=23.1.0
